requirements.txt
fastapi
sentence-transformers
uvicorn
torch
python-dotenv
